requirements.txt
asgiref==3.8.1
Django==4.2.17
django-debug-toolbar==5.0.1
django-filter==24.3
djangorestframework==3.15.2
Markdown==3.7
pillow==11.1.0
psycopg==3.2.4
psycopg-binary==3.2.4
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.12.2
